+++ gcloud/core/views.py
@@ -24,11 +24,10 @@ from django.shortcuts import render
 
 from blueapps.account.middlewares import LoginRequiredMiddleware
 
-from bk_api import is_user_functor, is_user_auditor
-
 from gcloud import exceptions
 from gcloud.conf import settings
 from gcloud.core.models import UserBusiness
+from gcloud.core.api_adapter import is_user_functor, is_user_auditor
 from gcloud.core.utils import prepare_user_business
 
 logger = logging.getLogger("root")
